git commit

+++ fe.cpp
@@ -6,6 +6,6 @@ int f(x);
 int main()
 {
     int x=2;
-    double res = f(x);
+    int res = f(x);
     return 0;
 }
